+++ app/Http/Controllers/Admin/ProductController.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Http\Controllers\Admin;
+
+use App\Http\Controllers\Controller;
+
+class ProductController extends Controller
+{
+    //
+}
