+++ app/Http/Controllers/Admin/PurchaseController.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Http\Controllers\Admin;
+
+use App\Http\Controllers\Controller;
+
+class PurchaseController extends Controller
+{
+    //
+}
