--- app/Http/Controllers/PurchaseController.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-
-class PurchaseController extends Controller
-{
-    //
-}
