--- app/Http/Controllers/SaleDetailController.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-
-class SaleDetailController extends Controller
-{
-    //
-}
