+++ resources/views/admin/user_types/index.blade.php
@@ -1,23 +1,26 @@
-@extends('layouts.app')
+@extends('app')
 @section('content')
     <div class="pagetitle">
-        <h1>Departamentos</h1>
+        <h1>Tipos de Usuarios</h1>
         <nav>
             <ol class="breadcrumb">
-                <li class="breadcrumb-item">Departamentos</li>
+                <li class="breadcrumb-item">Usuarios</li>
+                <li class="breadcrumb-item active">Tipos de Usuarios</li>
             </ol>
         </nav>
     </div>
     <div class=" card">
         <div class="card-body">
             <div style="display: flex; justify-content: right;" class="py-3">
-                <button onclick="showCreate()" class="btn btn-primary rounded-3"><i class="fa-solid fa-plus"></i> Crear</button>
+                <button onclick="showCreate()" class="btn btn-primary rounded-3"><i class="fa-solid fa-plus"></i> Crear
+                </button>
             </div>
-            <table id="department-types-table" class="table table-bordered">
+            <table id="user-types-table" class="table table-bordered">
                 <thead>
                 <tr>
                     <th scope="col">#</th>
-                    <th scope="col">Nombre</th>
+                    <th scope="col">Descripción</th>
+                    <th scope="col">Estado</th>
                     <th scope="col">Acciones</th>
                 </tr>
                 </thead>
@@ -38,11 +41,18 @@
                         @method('PUT')
                         @csrf
                         <div class="col-md-4 position-relative">
-                            <h6 for="validationTooltip01" class="form-label">Nombre</h6>
-                            <input name="name" type="text" class="form-control" id="edit-name" required>
+                            <h6 for="validationTooltip01" class="form-label">Descripción</h6>
+                            <input name="description" type="text" class="form-control" id="edit-description" required>
                             <div class="invalid-feedback">
                                 Este campo es obligatorio.
                             </div>
+                        </div>
+                        <div class="col-md-3">
+                            <h6 for="validationDefault04" class="form-label">Estado</h6>
+                            <select name="active" class="form-select" id="edit-status" required>
+                                <option value="1">Activo</option>
+                                <option value="0">Inactivo</option>
+                            </select>
                         </div>
                     </form>
                 </div>
@@ -63,7 +73,8 @@
                     <button type="button" class="btn fw-bold" data-bs-dismiss="modal" aria-label="Close">x</button>
                 </div>
                 <div class="modal-body align-content-center">
-                    <form class="row g-3 needs-validation" action="{{ route('user-types.store') }}" method="POST" novalidate id="form-create">
+                    <form class="row g-3 needs-validation" action="{{ route('user-types.store') }}" method="POST"
+                          novalidate id="form-create">
                         @csrf
                         <div class="col-md-4 position-relative">
                             <h6 for="validationTooltip01" class="form-label">Descripción</h6>
@@ -91,39 +102,35 @@
 @endsection
 @push('scripts')
     <script>
-        let table = $('#department-types-table')
+        let table = $('#user-types-table')
         table.DataTable({
             "language": {
-                "url": "//cdn.datatables.net/plug-ins/1.10.15/i18n/Spanish.json"},
+                "url": "//cdn.datatables.net/plug-ins/1.10.15/i18n/Spanish.json"
+            },
             destroy: true,
             responsive: true,
             processing: true,
             dom: 'Bfrtip',
             buttons: [{
                 extend: 'excelHtml5',
-                title: 'Departamentos',
-                filename: 'Departamentos',
+                title: 'Tipos de usuarios',
+                filename: 'tipos_de_usuarios',
             }],
-            ajax: '/departments-list',
+            ajax: '/user-types-list',
             columns: [
-                { data: 'id'},
-                { data: 'name' },
-                {
-                    data: 'id',
-                    render: function (id) {
-                        return '<button onclick="showModalEdit('+ id +')" type="button" class="btn btn-warning" title="Editar"><i class="fa-sharp fa-solid fa-pen-to-square"></i></button>'+
-                            ' <button onclick="destroy('+"'"+"departments/"+ id +"'"+')" type="button" class="btn btn-danger" title="Eliminar"><i class="fa-solid fa-trash"></i></button>'
-                    }
-                },
+                {data: 'id'},
+                {data: 'description'},
+                {data: 'active'},
+                {data: 'actions'},
             ],
         });
 
         function showModalEdit(id) {
-            axios.get('/user-types/'+id)
+            axios.get('/user-types/' + id)
                 .then(function (response) {
                     document.getElementById('edit-description').value = response.data.description
                     document.getElementById('edit-status').value = response.data.active
-                    document.getElementById('form-edit').setAttribute('action', '/user-types/'+id)
+                    document.getElementById('form-edit').setAttribute('action', '/user-types/' + id)
                 })
                 .catch(function (error) {
                     showAlert('error', error.data.message)
