+++ resources/views/auth/auth.blade.php
