+++ routes/web.php
@@ -14,5 +14,5 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::get('/', function () {
-    return view('app.dashboard.dashboard');
+    return view('admin.dashboard.dashboard');
 });
